+++ src/routes/index.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
 	import PostCard from '$lib/post/PostCard.svelte';
 	import Pagination from '$lib/pagination/Pagination.svelte';
-	import AuthorsListCard from '$lib/author/AuthorsListCard.svelte';
-	import CategoriesListCard from '$lib/category/CategoriesListCard.svelte';
+	import AuthorListCard from '$lib/author/AuthorListCard.svelte';
+	import CategoryListCard from '$lib/category/CategoryListCard.svelte';
 	import type { Post } from '$types/post';
 	import type { User } from '$types/user';
 	import type { Category } from '$types/category';
@@ -73,14 +73,14 @@
 			<div class="px-8">
 				<h1 class="mb-4 text-xl font-bold text-gray-700">Authors</h1>
 				<div class="max-w-sm mx-auto">
-					<AuthorsListCard {authors} />
+					<AuthorListCard {authors} />
 				</div>
 			</div>
 			<div class="px-8 mt-10">
 				<h1 class="mb-4 text-xl font-bold text-gray-700">Categories</h1>
 			</div>
 			<div class="max-w-sm mx-auto">
-				<CategoriesListCard {categories} />
+				<CategoryListCard {categories} />
 			</div>
 		</div>
 	</div>
